Reorganize so what we have is `hakn` serving.
Rather than going full-on `mink` and combining everything, this refactors things so that what we have so far is just a shrink-wrapped version of the parts we want for Knative Serving.

We stick with the `knative-serving` namespace, so stuff like the upstream monitoring installs cleanly.

The expectation is that next we'll do `knative-eventing` and we can explore combining things in the future, but this will already improve things be being HA and reducing the number of pods we're running on KinD.